AVX-32997: Refactor CSP calls
