+++ csp/sg.py
@@ -1,3 +1,5 @@
+import os
+
 import botocore
 
 from errors.exceptions import AvxError
@@ -40,3 +42,39 @@ def temp_add_security_group_access(client, controller_instanceobj, api_private_a
         print(str(err))
         raise
     return False, sgs[0]
+
+
+def create_new_sg(client):
+    """ Creates a new security group"""
+    instance_name = os.environ.get('AVIATRIX_TAG')
+    vpc_id = os.environ.get('VPC_ID')
+    try:
+        resp = client.create_security_group(Description='Aviatrix Controller',
+                                            GroupName=instance_name,
+                                            VpcId=vpc_id)
+        sg_id = resp['GroupId']
+    except (botocore.exceptions.ClientError, KeyError) as err:
+        if "InvalidGroup.Duplicate" in str(err):
+            rsp = client.describe_security_groups(GroupNames=[instance_name])
+            sg_id = rsp['SecurityGroups'][0]['GroupId']
+        else:
+            raise AvxError(str(err)) from err
+    try:
+        client.authorize_security_group_ingress(
+            GroupId=sg_id,
+            IpPermissions=[
+                {'IpProtocol': 'tcp',
+                 'FromPort': 443,
+                 'ToPort': 443,
+                 'IpRanges': [{'CidrIp': '0.0.0.0/0'}]},
+                {'IpProtocol': 'tcp',
+                 'FromPort': 80,
+                 'ToPort': 80,
+                 'IpRanges': [{'CidrIp': '0.0.0.0/0'}]}
+            ])
+    except botocore.exceptions.ClientError as err:
+        if "InvalidGroup.Duplicate" in str(err) or "InvalidPermission.Duplicate" in str(err):
+            pass
+        else:
+            raise AvxError(str(err)) from err
+    return sg_id
